add fragments
