+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,7 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
-    <style name="Theme.AppCompat.DayNight.NoActionBar." parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+<!--    <style name="Theme.AppCompat.DayNight.NoActionBar." parent="Theme.MaterialComponents.DayNight.DarkActionBar">-->
+    <style name="AppTheme" parent="Theme.MaterialComponents.Light.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/purple_200</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
